refactor of  msg_server and msg_server_test and minor fix
